+++ src/library/String.cpp
@@ -16,7 +16,7 @@ void String::init(stack *st)
 
     addFunctionL(object::stringPrototype, n::readOperator, {
         argsConvertibleGuard<number>(args);
-        int pos = static_cast<int>(args[0]->getConverted<number>());
+        int pos = static_cast<int>(args[0].getConverted<number>());
         const std::string &me = thisObj->get<const std::string>();
         assert<std::greater_equal>(pos, 0);
         assert<std::less>(pos, me.size());
@@ -25,7 +25,7 @@ void String::init(stack *st)
 
     addFunctionL(object::stringPrototype, "getByteAt"_n, {
         argsConvertibleGuard<number>(args);
-        int pos = static_cast<int>(args[0]->getConverted<number>());
+        int pos = static_cast<int>(args[0].getConverted<number>());
         const std::string &me = thisObj->get<const std::string>();
         assert<std::greater_equal>(pos, 0);
         assert<std::less>(pos, me.size());
@@ -34,8 +34,8 @@ void String::init(stack *st)
 
     addFunctionL(object::stringPrototype, "setByteAt"_n, {
         argsConvertibleGuard<number, number>(args);
-        int pos = static_cast<int>(args[0]->getConverted<number>());
-        int byte = static_cast<int>(args[1]->getConverted<number>());
+        int pos = static_cast<int>(args[0].getConverted<number>());
+        int byte = static_cast<int>(args[1].getConverted<number>());
         std::string &me = thisObj->get<std::string>();
         assert<std::greater_equal>(pos, 0);
         assert<std::less>(pos, me.size());
@@ -51,18 +51,18 @@ void String::init(stack *st)
         argsConvertibleGuard<number, std::string>(args);
         std::string &dest = thisObj->get<std::string>();
         const std::string &src = args[1]->get<const std::string>();
-        int destPos = static_cast<int>(args[0]->getConverted<number>());
+        int destPos = static_cast<int>(args[0].getConverted<number>());
         int srcPos = 0;
         int srcLength = src.size();
         if (args.size() > 2)
         {
             argsConvertibleGuard<nullptr_t, nullptr_t, number>(args);
-            srcPos = static_cast<int>(args[2]->getConverted<number>());
+            srcPos = static_cast<int>(args[2].getConverted<number>());
             srcLength = srcLength - srcPos;
             if (args.size() > 3)
             {
                 argsConvertibleGuard<nullptr_t, nullptr_t, number, number>(args);
-                srcLength = std::min(srcLength, static_cast<int>(args[3]->getConverted<number>()));
+                srcLength = std::min(srcLength, static_cast<int>(args[3].getConverted<number>()));
             }
         }
         assert<std::greater_equal>(destPos, 0);
@@ -83,12 +83,12 @@ void String::init(stack *st)
         if (args.size() > 1)
         {
             argsConvertibleGuard<nullptr_t, number>(args);
-            srcPos = static_cast<int>(args[1]->getConverted<number>());
+            srcPos = static_cast<int>(args[1].getConverted<number>());
             srcLength = srcLength - srcPos;
             if (args.size() > 2)
             {
                 argsConvertibleGuard<nullptr_t, number, number>(args);
-                srcLength = std::min(srcLength, static_cast<int>(args[2]->getConverted<number>()));
+                srcLength = std::min(srcLength, static_cast<int>(args[2].getConverted<number>()));
             }
         }
         assert<std::greater_equal>(srcPos, 0);
@@ -101,12 +101,12 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "substring"_n, {
         argsConvertibleGuard<number>(args);
         const std::string &me = thisObj->get<const std::string>();
-        int pos = static_cast<int>(args[0]->getConverted<number>());
+        int pos = static_cast<int>(args[0].getConverted<number>());
         int length = me.size() - pos;
         if (args.size() > 1)
         {
             argsConvertibleGuard<nullptr_t, number>(args);
-            length = std::min(length, static_cast<int>(args[1]->getConverted<number>()));
+            length = std::min(length, static_cast<int>(args[1].getConverted<number>()));
         }
         assert<std::greater_equal>(pos, 0);
         assert<std::greater_equal>(length, 0);
@@ -117,12 +117,12 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "erase"_n, {
         argsConvertibleGuard<number>(args);
         std::string &me = thisObj->get<std::string>();
-        int pos = static_cast<int>(args[0]->getConverted<number>());
+        int pos = static_cast<int>(args[0].getConverted<number>());
         int length = me.size() - pos;
         if (args.size() > 1)
         {
             argsConvertibleGuard<nullptr_t, number>(args);
-            length = std::min(length, static_cast<int>(args[1]->getConverted<number>()));
+            length = std::min(length, static_cast<int>(args[1].getConverted<number>()));
         }
         assert<std::greater_equal>(pos, 0);
         assert<std::greater_equal>(length, 0);
@@ -134,7 +134,7 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "split"_n, {
         argsConvertibleGuard<std::string>(args);
         const std::string &me = thisObj->get<const std::string>();
-        const std::string &delim = args[0]->getConverted<const std::string>();
+        const std::string &delim = args[0].getConverted<const std::string>();
         int begin = 0, end, delimSize = delim.size();
 
         object::type::Array res;
@@ -151,10 +151,10 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "indexOf"_n, {
         argsConvertibleGuard<std::string>(args);
         std::string &me = thisObj->get<std::string>();
-        std::string key = args[0]->getConverted<std::string>();
+        std::string key = args[0].getConverted<std::string>();
         int pos = 0;
         if (args.size() >= 2)
-            pos = static_cast<int>(args[1]->getConverted<number>());
+            pos = static_cast<int>(args[1].getConverted<number>());
         assert<std::greater_equal>(pos, 0);
         assert<std::less_equal>(pos, me.size());
         return object::makeObject(static_cast<number>(me.find(key, pos)));
@@ -163,15 +163,15 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "search"_n, {
         argsConvertibleGuard<std::string>(args);
         const std::string &me = thisObj->get<const std::string>();
-        std::regex regex(args[0]->getConverted<const std::string>());
+        std::regex regex(args[0].getConverted<const std::string>());
         std::smatch matchResults;
         int pos = 0;
         int length = me.size();
 
         if (args.size() >= 2)
-            pos = static_cast<int>(args[1]->getConverted<number>());
+            pos = static_cast<int>(args[1].getConverted<number>());
         if (args.size() >= 3)
-            length = static_cast<int>(args[2]->getConverted<number>());
+            length = static_cast<int>(args[2].getConverted<number>());
 
         std::regex_search(me.begin() + pos, me.begin() + pos + length, matchResults, regex);
         object::type::Array array(matchResults.size());
@@ -187,14 +187,14 @@ void String::init(stack *st)
     addFunctionL(object::stringPrototype, "searchAll"_n, {
         argsConvertibleGuard<std::string>(args);
         const std::string &me = thisObj->get<const std::string>();
-        std::regex regex(args[0]->getConverted<const std::string>());
+        std::regex regex(args[0].getConverted<const std::string>());
         int pos = 0;
         int length = me.size();
 
         if (args.size() >= 2)
-            pos = static_cast<int>(args[1]->getConverted<number>());
+            pos = static_cast<int>(args[1].getConverted<number>());
         if (args.size() >= 3)
-            length = static_cast<int>(args[2]->getConverted<number>());
+            length = static_cast<int>(args[2].getConverted<number>());
 
         auto begin = std::sregex_iterator(me.begin() + pos, me.begin() + pos + length, regex);
         auto end = std::sregex_iterator();