+++ src/library/Tcp.cpp
@@ -17,8 +17,8 @@ void Tcp::init(stack *st)
     addFunctionL(tcpClientPrototype, "connect"_n, {
         argsConvertibleGuard<std::string, number>(args);
 
-        std::string adress = args[0]->getConverted<std::string>();
-        int port = static_cast<int>(args[1]->getConverted<number>());
+        std::string adress = args[0].getConverted<std::string>();
+        int port = static_cast<int>(args[1].getConverted<number>());
 
         return object::makeObject(coroutine<object::objectPtr>::makeCoroutine([adress, port, thisObj]() {
             auto t = thisObj->get<std::shared_ptr<tcpClient>>();
@@ -45,7 +45,7 @@ void Tcp::init(stack *st)
     addFunctionL(tcpClientPrototype, "send"_n, {
         argsConvertibleGuard<std::string>(args);
 
-        std::string str = args[0]->getConverted<std::string>();
+        std::string str = args[0].getConverted<std::string>();
 
         return object::makeObject(coroutine<object::objectPtr>::makeCoroutine([str, thisObj]() {
             auto t = thisObj->get<std::shared_ptr<tcpClient>>();
@@ -82,7 +82,7 @@ void Tcp::init(stack *st)
     addFunctionL(tcpServerPrototype, "bind"_n, {
         argsConvertibleGuard<number>(args);
 
-        int port = static_cast<int>(args[0]->getConverted<number>());
+        int port = static_cast<int>(args[0].getConverted<number>());
 
         return object::makeObject(coroutine<object::objectPtr>::makeCoroutine([port, thisObj]() {
             auto t = thisObj->get<std::shared_ptr<tcpServer>>();