+++ easypt_vsce/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+### 3.0.5 - 3.0.2
+
+- Fix number highlights.
+
+### 3.0.1
+
+- New number highlights added.
+
 ### 3.0.0
 
 - Auto-completion added.